third_party/WebKit/LayoutTests/editing/spelling/cold_mode_type_idle.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script src="spellcheck_test.js"></script>
<script>
spellcheck_test(
    '<div contenteditable>|</div>',
    'insertText zz',
    '<div contenteditable>zz</div>',
    {
      title: 'No immediate checking during typing',
      callback: sample => spellcheck_test(
          sample,
          '',
          '<div contenteditable>#zz#</div>',
          {
            title: 'Text is eventually checked in cold mode',
            needsFullCheck: true
          })
    });
</script>
