third_party/WebKit/LayoutTests/platform/win/editing/spelling/cold_mode_type_idle-expected.txt

This is a testharness.js-based test.
PASS No immediate checking during typing 
PASS Text is eventually checked in cold mode 
Harness: the test ran to completion.
